Extract response handling from Events::handleEditor()
