+++ libraries/classes/Database/Events.php
@@ -21,7 +21,6 @@ use function htmlspecialchars;
 use function in_array;
 use function intval;
 use function is_string;
-use function mb_strtoupper;
 use function sprintf;
 use function str_contains;
 use function strtoupper;
@@ -71,185 +70,93 @@ class Events
     /**
      * Handles editor requests for adding or editing an item
      */
-    public function handleEditor(): void
+    public function handleEditor(): string
     {
-        $GLOBALS['errors'] ??= null;
-        $GLOBALS['message'] ??= null;
+        $sqlQuery = '';
 
-        if (! empty($_POST['editor_process_add']) || ! empty($_POST['editor_process_edit'])) {
-            $sqlQuery = '';
+        $itemQuery = $this->getQueryFromRequest();
 
-            $itemQuery = $this->getQueryFromRequest();
-
-            // set by getQueryFromRequest()
-            if (! count($GLOBALS['errors'])) {
-                // Execute the created query
-                if (! empty($_POST['editor_process_edit'])) {
-                    // Backup the old trigger, in case something goes wrong
-                    $createItem = self::getDefinition($this->dbi, $GLOBALS['db'], $_POST['item_original_name']);
-                    $dropItem = 'DROP EVENT IF EXISTS '
-                        . Util::backquote($_POST['item_original_name'])
-                        . ";\n";
-                    $result = $this->dbi->tryQuery($dropItem);
-                    if (! $result) {
-                        $GLOBALS['errors'][] = sprintf(
-                            __('The following query has failed: "%s"'),
-                            htmlspecialchars($dropItem),
-                        )
+        // set by getQueryFromRequest()
+        if (! count($GLOBALS['errors'])) {
+            // Execute the created query
+            if (! empty($_POST['editor_process_edit'])) {
+                // Backup the old trigger, in case something goes wrong
+                $createItem = self::getDefinition($this->dbi, $GLOBALS['db'], $_POST['item_original_name']);
+                $dropItem = 'DROP EVENT IF EXISTS '
+                    . Util::backquote($_POST['item_original_name'])
+                    . ";\n";
+                $result = $this->dbi->tryQuery($dropItem);
+                if (! $result) {
+                    $GLOBALS['errors'][] = sprintf(
+                        __('The following query has failed: "%s"'),
+                        htmlspecialchars($dropItem),
+                    )
                         . '<br>'
                         . __('MySQL said: ') . $this->dbi->getError();
-                    } else {
-                        $result = $this->dbi->tryQuery($itemQuery);
-                        if (! $result) {
-                            $GLOBALS['errors'][] = sprintf(
-                                __('The following query has failed: "%s"'),
-                                htmlspecialchars($itemQuery),
-                            )
-                            . '<br>'
-                            . __('MySQL said: ') . $this->dbi->getError();
-                            // We dropped the old item, but were unable to create
-                            // the new one. Try to restore the backup query
-                            $result = $this->dbi->tryQuery($createItem);
-                            if (! $result) {
-                                $GLOBALS['errors'] = $this->checkResult($createItem, $GLOBALS['errors']);
-                            }
-                        } else {
-                            $GLOBALS['message'] = Message::success(
-                                __('Event %1$s has been modified.'),
-                            );
-                            $GLOBALS['message']->addParam(
-                                Util::backquote($_POST['item_name']),
-                            );
-                            $sqlQuery = $dropItem . $itemQuery;
-                        }
-                    }
                 } else {
-                    // 'Add a new item' mode
                     $result = $this->dbi->tryQuery($itemQuery);
                     if (! $result) {
                         $GLOBALS['errors'][] = sprintf(
                             __('The following query has failed: "%s"'),
                             htmlspecialchars($itemQuery),
                         )
-                        . '<br><br>'
-                        . __('MySQL said: ') . $this->dbi->getError();
+                            . '<br>'
+                            . __('MySQL said: ') . $this->dbi->getError();
+                        // We dropped the old item, but were unable to create
+                        // the new one. Try to restore the backup query
+                        $result = $this->dbi->tryQuery($createItem);
+                        if (! $result) {
+                            $GLOBALS['errors'] = $this->checkResult($createItem, $GLOBALS['errors']);
+                        }
                     } else {
                         $GLOBALS['message'] = Message::success(
-                            __('Event %1$s has been created.'),
+                            __('Event %1$s has been modified.'),
                         );
                         $GLOBALS['message']->addParam(
                             Util::backquote($_POST['item_name']),
                         );
-                        $sqlQuery = $itemQuery;
+                        $sqlQuery = $dropItem . $itemQuery;
                     }
-                }
-            }
-
-            if (count($GLOBALS['errors'])) {
-                $GLOBALS['message'] = Message::error(
-                    '<b>'
-                    . __(
-                        'One or more errors have occurred while processing your request:',
-                    )
-                    . '</b>',
-                );
-                $GLOBALS['message']->addHtml('<ul>');
-                foreach ($GLOBALS['errors'] as $string) {
-                    $GLOBALS['message']->addHtml('<li>' . $string . '</li>');
-                }
-
-                $GLOBALS['message']->addHtml('</ul>');
-            }
-
-            $output = Generator::getMessage($GLOBALS['message'], $sqlQuery);
-
-            if ($this->response->isAjax()) {
-                if ($GLOBALS['message']->isSuccess()) {
-                    $events = $this->getDetails($GLOBALS['db'], $_POST['item_name']);
-                    $event = $events[0];
-                    $this->response->addJSON(
-                        'name',
-                        htmlspecialchars(
-                            mb_strtoupper($_POST['item_name']),
-                        ),
-                    );
-                    if (! empty($event)) {
-                        $sqlDrop = sprintf(
-                            'DROP EVENT IF EXISTS %s',
-                            Util::backquote($event['name']),
-                        );
-                        $this->response->addJSON(
-                            'new_row',
-                            $this->template->render('database/events/row', [
-                                'db' => $GLOBALS['db'],
-                                'table' => $GLOBALS['table'],
-                                'event' => $event,
-                                'has_privilege' => Util::currentUserHasPrivilege('EVENT', $GLOBALS['db']),
-                                'sql_drop' => $sqlDrop,
-                                'row_class' => '',
-                            ]),
-                        );
-                    }
-
-                    $this->response->addJSON('insert', ! empty($event));
-                    $this->response->addJSON('message', $output);
-                } else {
-                    $this->response->setRequestStatus(false);
-                    $this->response->addJSON('message', $GLOBALS['message']);
-                }
-
-                $this->response->addJSON('tableType', 'events');
-                $this->response->callExit();
-            }
-        }
-
-        /**
-         * Display a form used to add/edit a trigger, if necessary
-         */
-        if (
-            ! count($GLOBALS['errors'])
-            && (! empty($_POST['editor_process_add'])
-            || ! empty($_POST['editor_process_edit'])
-            || (empty($_REQUEST['add_item'])
-            && empty($_REQUEST['edit_item'])
-            && empty($_POST['item_changetype'])))
-        ) {
-            return;
-        }
-
-        // FIXME: this must be simpler than that
-        $operation = '';
-        $title = '';
-        $item = null;
-        $mode = '';
-        if (! empty($_POST['item_changetype'])) {
-            $operation = 'change';
-        }
-
-        // Get the data for the form (if any)
-        if (! empty($_REQUEST['add_item'])) {
-            $title = __('Add event');
-            $item = $this->getDataFromRequest();
-            $mode = 'add';
-        } elseif (! empty($_REQUEST['edit_item'])) {
-            $title = __('Edit event');
-            if (
-                ! empty($_REQUEST['item_name'])
-                && empty($_POST['editor_process_edit'])
-                && empty($_POST['item_changetype'])
-            ) {
-                $item = $this->getDataFromName($_REQUEST['item_name']);
-                if ($item !== null) {
-                    $item['item_original_name'] = $item['item_name'];
                 }
             } else {
-                $item = $this->getDataFromRequest();
+                // 'Add a new item' mode
+                $result = $this->dbi->tryQuery($itemQuery);
+                if (! $result) {
+                    $GLOBALS['errors'][] = sprintf(
+                        __('The following query has failed: "%s"'),
+                        htmlspecialchars($itemQuery),
+                    )
+                        . '<br><br>'
+                        . __('MySQL said: ') . $this->dbi->getError();
+                } else {
+                    $GLOBALS['message'] = Message::success(
+                        __('Event %1$s has been created.'),
+                    );
+                    $GLOBALS['message']->addParam(
+                        Util::backquote($_POST['item_name']),
+                    );
+                    $sqlQuery = $itemQuery;
+                }
             }
-
-            $mode = 'edit';
         }
 
-        $this->sendEditor($mode, $item, $title, $GLOBALS['db'], $operation);
+        if (count($GLOBALS['errors'])) {
+            $GLOBALS['message'] = Message::error(
+                '<b>'
+                . __(
+                    'One or more errors have occurred while processing your request:',
+                )
+                . '</b>',
+            );
+            $GLOBALS['message']->addHtml('<ul>');
+            foreach ($GLOBALS['errors'] as $string) {
+                $GLOBALS['message']->addHtml('<li>' . $string . '</li>');
+            }
+
+            $GLOBALS['message']->addHtml('</ul>');
+        }
+
+        return Generator::getMessage($GLOBALS['message'], $sqlQuery);
     }
 
     /**
@@ -495,7 +402,7 @@ class Events
      * @param string       $db        Database
      * @param string       $operation Operation 'change' or ''
      */
-    private function sendEditor(string $mode, array|null $item, string $title, string $db, string $operation): void
+    public function sendEditor(string $mode, array|null $item, string $title, string $db, string $operation): void
     {
         if ($item !== null) {
             $editor = $this->getEditorForm($mode, $operation, $item);
